backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pymupdf==1.24.14
python-docx==1.1.2
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.57.2
chromadb==0.5.23
python-multipart==0.0.18
